+++ frontend/src/metabase/query_builder/components/DataSelector/DataSelector.jsx
@@ -20,7 +20,6 @@ import LoadingAndErrorWrapper from "metabase/components/LoadingAndErrorWrapper";
 
 import MetabaseSettings from "metabase/lib/settings";
 import { getSchemaName } from "metabase/lib/schema";
-import { isSyncCompleted } from "metabase/lib/syncing";
 import { isDatabaseWritebackEnabled } from "metabase/writeback/utils";
 
 import Databases from "metabase/entities/databases";
@@ -36,12 +35,12 @@ import {
 } from "./data-search";
 import SavedQuestionPicker from "./saved-question-picker/SavedQuestionPicker";
 import DataSelectorLoading from "./DataSelectorLoading";
+import DatabaseSchemaPicker from "./DataSelectorDatabaseSchemaPicker";
 import FieldPicker from "./DataSelectorFieldPicker";
 import TablePicker from "./DataSelectorTablePicker";
 import {
   DataBucketList,
   DataBucketListItem,
-  PickerSpinner,
   RawDataBackButton,
   CollectionDatasetSelectList,
   CollectionDatasetAllDataLink,
@@ -1312,95 +1311,5 @@ const SchemaPicker = ({
         showItemArrows={hasNextStep}
       />
     </div>
-  );
-};
-
-const DatabaseSchemaPicker = ({
-  databases,
-  selectedDatabase,
-  selectedSchema,
-  onChangeSchema,
-  onChangeDatabase,
-  hasNextStep,
-  isLoading,
-  hasBackButton,
-  onBack,
-  hasInitialFocus,
-}) => {
-  if (databases.length === 0) {
-    return <DataSelectorLoading />;
-  }
-
-  const sections = databases.map(database => ({
-    name: database.is_saved_questions ? t`Saved Questions` : database.name,
-    items:
-      !database.is_saved_questions && database.schemas.length > 1
-        ? database.schemas.map(schema => ({
-            schema,
-            name: schema.displayName(),
-          }))
-        : [],
-    className: database.is_saved_questions ? "bg-light" : null,
-    icon: database.is_saved_questions ? "all" : "database",
-    loading:
-      selectedDatabase?.id === database.id &&
-      database.schemas.length === 0 &&
-      isLoading,
-    active: database.is_saved_questions || isSyncCompleted(database),
-  }));
-
-  if (hasBackButton) {
-    sections.unshift({
-      name: <RawDataBackButton />,
-      active: true,
-    });
-  }
-
-  let openSection = selectedSchema
-    ? databases.findIndex(db => db.id === selectedSchema.database.id)
-    : selectedDatabase
-    ? databases.findIndex(db => db.id === selectedDatabase.id)
-    : -1;
-
-  if (openSection >= 0 && databases[openSection]?.schemas.length === 1) {
-    openSection = -1;
-  }
-
-  return (
-    <AccordionList
-      id="DatabaseSchemaPicker"
-      key="databaseSchemaPicker"
-      className="text-brand"
-      hasInitialFocus={hasInitialFocus}
-      sections={sections}
-      onChange={item => onChangeSchema(item.schema)}
-      onChangeSection={(_section, sectionIndex) => {
-        const isNavigationSection = hasBackButton && sectionIndex === 0;
-        if (isNavigationSection) {
-          onBack();
-          return false;
-        }
-        // the "go back" button is also a section,
-        // so need to take its index in mind
-        const database = hasBackButton
-          ? databases[sectionIndex - 1]
-          : databases[sectionIndex];
-        onChangeDatabase(database);
-        return true;
-      }}
-      itemIsSelected={schema => schema === selectedSchema}
-      renderSectionIcon={item =>
-        item.icon && (
-          <Icon className="Icon text-default" name={item.icon} size={18} />
-        )
-      }
-      renderSectionExtra={item =>
-        !item.active && <PickerSpinner size={16} borderWidth={2} />
-      }
-      renderItemIcon={() => <Icon name="folder" size={16} />}
-      initiallyOpenSection={openSection}
-      alwaysTogglable={true}
-      showItemArrows={hasNextStep}
-    />
   );
 };
