+++ frontend/src/metabase/query_builder/components/DataSelector/DataSelectorDatabaseSchemaPicker/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./DataSelectorDatabaseSchemaPicker";
